Cleanup and Bugfixes
Joystick death bug fix
Commenting
Some cleanup
Added some functions to call
