+++ Assets/AnimationTesting.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: eb8037d687438e342be4d52f6ed65012
+timeCreated: 1519252870
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
