+++ Assets/Prefabs/Obstacles/Explosions/SmallExplosion.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 4b19f37fc37135842a919a63845f9e9f
+timeCreated: 1518822341
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
