--- Assets/Scripts/DestroyThis.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: cc650e3ff95cc6e459dca384a10eacf6
-timeCreated: 1518375738
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
