--- Assets/Scripts/ExploderObstacle.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 5a9f25d0046196948978cc088a0f0c01
-timeCreated: 1518466487
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
